touches

+++ routes/admin/shop.js
@@ -3,8 +3,8 @@ const shopController = require('../../controllers/shop');
 const express = require('express');
 const { check, validationResult } = require('express-validator/check');
 const router = express.Router();
-//const multer = require('multer'); //Multipart form data (files)
 const upload = require('../../middleware/upload');
+const fs = require('fs');
 
 router
     .get('/', async (request, response) => {
@@ -53,22 +53,31 @@ router
 
     //Edit category
     .get('/categories/:id', async (request, response) => {
-        response.locals.metaTags = {
-            title: 'Admin - Shop - Categories - edit',
-            description: 'Here goes the description',
-            keywords: 'Here goes keywords'
-        };
-        response.render('admin/shop/category', {
-            layout: 'admin',
-            success: request.session.success,
-            errors: request.session.errors,
-            inputs: request.session.inputs,
-            category: await shopController.findCategory(request.params.id),
-        });
+        try {
+            const category = await shopController.findCategory(request.params.id);
+            if (!category) {
+                throw new Error('Kategorien blev ikke fundet');
+            }
+            response.locals.metaTags = {
+                title: 'Admin - Shop - Categories - edit',
+                description: 'Here goes the description',
+                keywords: 'Here goes keywords'
+            };
+            response.render('admin/shop/category', {
+                layout: 'admin',
+                success: request.session.success,
+                errors: request.session.errors,
+                inputs: request.session.inputs,
+                category: category,
+            });
 
-        request.session.success = null;
-        request.session.errors = null;
-        request.session.inputs = null;
+            request.session.success = null;
+            request.session.errors = null;
+            request.session.inputs = null;
+        } catch (err) {
+            response.render('error')
+        }
+
     })
 
     .post('/categories/:id', [
@@ -97,28 +106,33 @@ router
     })
 
     //Delete category
-    .delete('/categories/:id', async (request, response) => {
+    .post('/categories/:id/remove', async (request, response) => {
         try {
             const category = await shopController.findCategory(request.params.id);
             if (category) {
                 for (let product of category.products) {
                     await shopController.deleteProduct(product._id);
-                }
+                    if (product.imgPath !== '/uploads/default.png') {
+                        fs.unlink('public' + product.imgPath, function (err) {
+                            if (err) throw new Error(err);
+                            console.log(product.imgPath + 'Was deleted');
+                        });
+                    }
+                };
 
                 const result = await shopController.deleteCategory(request.params.id);
 
                 if (result) {
-                    response.sendStatus(200);
+                    request.session.success = { msg: 'Produktkategorien og tilhørende produkter blev slettet' };
+                    response.redirect('/admin/shop');
                 } else {
                     throw new Error('Der skete en fejl prøv igen senere')
                 }
             }
         } catch (err) {
-            request.session.errors = { msg: err };
-            response.redirect('/admin/shop/categories')
+            request.session.errors = { msg: 'what?' + err };
+            response.redirect('/admin/shop');
         }
-
-
     })
 
     //Create product
@@ -131,7 +145,7 @@ router
             }
         }),
         check('price', 'Pris er påkrævet').isDecimal(),
-        check('size', 'Størrelse er påkrævet').isString().isLength({min: 1})
+        check('size', 'Størrelse er påkrævet').isString().isLength({ min: 1 })
     ], async (request, response) => {
         console.log(request.file);
         const errors = validationResult(request);
@@ -140,8 +154,13 @@ router
             request.session.inputs = { name: request.body.name, price: request.body.price, size: request.body.size };
             response.redirect('/admin/shop');
         } else {
-            const { name, price, size, category, image } = request.body;
-            const product = await shopController.createProduct(name, price, size, '/uploads/' + request.file.filename);
+            const { name, price, size, category } = request.body;
+            let imgPath = '/uploads/default.png';
+            if (request.file) {
+                imgPath = '/uploads/' + request.file.filename;
+            }
+
+            const product = await shopController.createProduct(name, price, size, imgPath);
             const connect = await shopController.addProduktToCategory(category, product._id);
             if (product && connect) {
                 request.session.success = { msg: 'Success! - ' + name + ' er oprettet' };
@@ -156,6 +175,13 @@ router
 
     //Edit product
     .get('/categories/:id/products/:pid', async (request, response) => {
+        const product = await shopController.findProduct(request.params.pid);
+        const category = await shopController.findCategory(request.params.id);
+
+        if (!product && !category) {
+            response.send(404);
+        }
+
         response.locals.metaTags = {
             title: 'Admin - Shop - Edit product',
             description: 'Here goes the description',
@@ -165,8 +191,8 @@ router
             layout: 'admin',
             success: request.session.success,
             errors: request.session.errors,
-            product: await shopController.findProduct(request.params.pid),
-            category: await shopController.findCategory(request.params.id),
+            product: product,
+            category: category,
         });
 
         request.session.success = null;
@@ -174,7 +200,7 @@ router
         request.session.inputs = null;
     })
 
-    .post('/categories/:id/products/:pid', [
+    .post('/categories/:id/products/:pid', upload.single('productImage'), [
         check('name', 'Navn er påkrævet').isString().isLength({ min: 2 }).custom(async (name, { req }) => {
             if (req.body.name !== name) {
                 if (await shopController.checkProductName(name)) return Promise.reject('Dette navn er allerede i brug');
@@ -184,7 +210,7 @@ router
             }
         }),
         check('price').isDecimal(),
-        check('size').isString()
+        check('size').isString().isLength({ min: 2 })
     ], async (request, response) => {
         const errors = validationResult(request);
         if (!errors.isEmpty()) {
@@ -193,38 +219,65 @@ router
         } else {
             try {
                 const { category, name, price, size } = request.body;
-                const product = await shopController.updateProduct(request.params.pid, name, price, size);
+                const product = await shopController.findProduct(request.params.pid);
+                if (!product) {
+                    throw new Error('Produktet findes ikke')
+                }
 
-                if (product) {
+                let currentImage = product.imgPath;
+                if (request.file) {
+                    if (currentImage !== '/uploads/default.png') {
+                        fs.unlink('public' + currentImage, function (err) {
+                            if (err) throw new Error(err);
+                            console.log(currentImage + 'Was deleted')
+                        })
+                    }
+                    currentImage = '/uploads/' + request.file.filename;
+                }
+
+                const result = await shopController.updateProduct(request.params.pid, name, price, size, currentImage);
+
+                if (result) {
                     request.session.success = { msg: 'Success! - produktet er opdateret' };
                     response.redirect('/admin/shop/categories/' + request.params.id + '/products/' + request.params.pid);
                 } else {
                     throw new Error('Product update failed');
                 }
             } catch (err) {
-                request.session.errors = { msg: 'Ups der skete en fejl' };
+                request.session.errors = { msg: 'Ups der skete en fejl. ' + err };
                 response.redirect('/admin/shop/categories/' + request.params.id + '/products/' + request.params.pid);
             }
         }
     })
 
     //Delete product
-    .delete('categories/:id/products/:pid', async (request, response) => {
+    .post('/categories/:id/products/:pid/remove', async (request, response) => {
         try {
-            const product = await shopController.findProduct(request.params.id);
-            if (product) {
+            const product = await shopController.findProduct(request.params.pid);
+            const category = await shopController.findCategory(request.params.id)
+            if (product && category) {
                 const result = await shopController.deleteProduct(product._id);
-                const dc = await shopController.removeProduktFromCategory(request.params.id, product._id);
+                const dc = await shopController.removeProduktFromCategory(category._id, product._id);
+                const imgPath = result.imgPath;
 
                 if (result && dc) {
-                    response.sendStatus(200);
+                    if (imgPath !== '/uploads/default.png') {
+                        fs.unlink('public' + imgPath, function (err) {
+                            if (err) throw new Error(err);
+                            console.log(imgPath + 'Was deleted')
+                        });
+                    }
+
+                    if (request.url === '/categories/')
+                    request.session.success = { msg: 'Produktet blev slettet' };
+                    response.redirect('/admin/shop');
                 } else {
                     throw new Error('Der skete en fejl prøv igen senere')
                 }
             }
         } catch (err) {
-            request.session.errors = { msg: err };
-            response.redirect('/admin/shop');
+            request.session.errors = { msg: 'Noget gik galt: ' + err };
+            response.redirect('/categories/' + request.params.id + '/products/' + request.params.pid);
         }
     });
